+++ CHANGES.md
@@ -1,4 +1,9 @@
 
+## 5.1.0 (2023-12-21)
+
+* Simplify bounds calculation by using `TileMatrix.cellSize` instead of `TileMatrix.scaleDenominator`
+* remove `TileMatrixSet._resolution` private method
+
 ## 5.0.2 (2023-12-01)
 
 * Remove *alias* tiles in `.parent()`, `.children()`, `.neighbors()` and `.tiles()` methods for Variable Matrix Width TileMatrixSets (https://github.com/developmentseed/morecantile/pull/136)